// iped-webui/src/app/core/api/generated/model/facetResult.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {FacetBucket} from './facetBucket';


export interface FacetResult {
    field: string;
    buckets: Array<FacetBucket>;
}