// iped-webui/src/app/core/api/generated/model/filterClause.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FilterClause { 
    field: string;
    op: FilterClauseOpEnum;
    value?: any | null;
}
export enum FilterClauseOpEnum {
    eq = 'eq',
    ne = 'ne',
    in = 'in',
    gt = 'gt',
    gte = 'gte',
    lt = 'lt',
    lte = 'lte',
    contains = 'contains'
};