// iped-webui/src/app/core/api/generated/model/searchCreateRequest.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {PageRequest} from './pageRequest';
import {SortClause} from './sortClause';
import {FilterClause} from './filterClause';


export interface SearchCreateRequest {
    query: string;
    filters?: Array<FilterClause>;
    sort?: Array<SortClause>;
    page?: PageRequest;
}