// iped-webui/src/app/core/api/generated/model/searchResultsPage.ts
/**
 * IPED Web API - Phase 1 Initial Contracts
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {SearchResultItem} from './searchResultItem';
import {PageResponse} from './pageResponse';


export interface SearchResultsPage {
    searchId: string;
    total: number;
    page: PageResponse;
    items: Array<SearchResultItem>;
}